+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Manifesters now take an options object instead of just the Faker instance
+
 ## [0.3.0] - 2022-03-06
 
 ### Added